requirements.txt
Django~=5.1.1
django-debug-toolbar~=4.4.6
django-bootstrap5~=24.3 
django-bootstrap-icons~=0.8.6
djangorestframework~=3.14.0

